+++ src/tbox/network/ipv6.c
@@ -281,7 +281,7 @@ tb_bool_t tb_ipv6_cstr_set(tb_ipv6_ref_t ipv6, tb_char_t const* cstr)
                 // trace
                 tb_trace_d("scope_id: %u", temp.scope_id);
             }
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
             // is interface name?
             else 
             {