+++ src/tbox/platform/thread.c
@@ -49,7 +49,7 @@ typedef tb_pointer_t    tb_thread_retval_t;
 /* //////////////////////////////////////////////////////////////////////////////////////
  * private implementation
  */
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
 static tb_bool_t tb_thread_local_free(tb_iterator_ref_t iterator, tb_pointer_t item, tb_cpointer_t priv)
 {
     // the local
@@ -90,7 +90,7 @@ static tb_thread_retval_t tb_thread_func(tb_pointer_t priv)
         // call the thread function
         retval = (tb_thread_retval_t)(tb_size_t)func(args[1].ptr);
 
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
         // free all thread local data on the current thread
         tb_thread_local_walk(tb_thread_local_free, tb_null);
 #endif