+++ src/tbox/utils/trace.c
@@ -34,7 +34,7 @@
  */
 
 // the trace line maxn
-#ifdef TB_CONFIG_EMBED_ENABLE
+#ifdef TB_CONFIG_MICRO_ENABLE
 #       define TB_TRACE_LINE_MAXN       (256)
 #else
 #   ifdef __tb_small__
@@ -51,7 +51,7 @@
 // the mode
 static tb_size_t        g_mode = TB_TRACE_MODE_PRINT;
 
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
 // the file
 static tb_file_ref_t    g_file = tb_null;
 
@@ -85,7 +85,7 @@ tb_void_t tb_trace_exit()
     g_mode = TB_TRACE_MODE_PRINT;
 
     // clear file
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
     if (g_file && !g_bref) tb_file_exit(g_file);
     g_file = tb_null;
     g_bref = tb_false;
@@ -125,7 +125,7 @@ tb_bool_t tb_trace_mode_set(tb_size_t mode)
     // ok
     return tb_true;
 }
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
 tb_handle_t tb_trace_file()
 {
     // enter
@@ -205,7 +205,7 @@ tb_void_t tb_trace_done_with_args(tb_char_t const* prefix, tb_char_t const* modu
         tb_char_t*      e = g_line + sizeof(g_line);
 
         // print prefix to file
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
         if ((g_mode & TB_TRACE_MODE_FILE) && g_file) 
         {
             // print time to file
@@ -236,7 +236,7 @@ tb_void_t tb_trace_done_with_args(tb_char_t const* prefix, tb_char_t const* modu
         if (g_mode & TB_TRACE_MODE_PRINT) tb_print(b);
 
         // print it to file
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
         if ((g_mode & TB_TRACE_MODE_FILE) && g_file) 
         {
             // done
@@ -305,7 +305,7 @@ tb_void_t tb_trace_tail(tb_char_t const* format, ...)
         if (g_mode & TB_TRACE_MODE_PRINT) tb_print(g_line);
 
         // print it to file
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
         if ((g_mode & TB_TRACE_MODE_FILE) && g_file) 
         {
             // done
@@ -340,7 +340,7 @@ tb_void_t tb_trace_sync()
     if (g_mode & TB_TRACE_MODE_PRINT) tb_print_sync();
 
     // sync it to file
-#ifndef TB_CONFIG_EMBED_ENABLE
+#ifndef TB_CONFIG_MICRO_ENABLE
     if ((g_mode & TB_TRACE_MODE_FILE) && g_file) tb_file_sync(g_file);
 #endif